+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/JoinGame/JoinGameSessionController.cs
@@ -112,7 +112,7 @@ namespace Stormancer.Server.Plugins.Party.JoinGame
             {
                 var team = new Team { TeamId = tuple.team.TeamId };
                 var partyArg = new Models.Party { PartyId = partyId };
-                partyArg.Players.Add(session.User.Id, new Player { SessionId = session.SessionId.ToString(), UserId = session.User.Id });
+                partyArg.Players.Add(session.User.Id, new Player { SessionId = session.SessionId, UserId = session.User.Id });
                 team.Parties.Add(partyArg);
                 await gameSession.CreateReservationAsync(team, new Newtonsoft.Json.Linq.JObject { }, cancellationToken);